入力項目のロジックを使い回すためmixinに切り出し、共通部品をBase.vueに導入
